+++ PolyDeploy.DeployClient.Tests/InstallerTests.cs
@@ -19,7 +19,7 @@ namespace PolyDeploy.DeployClient.Tests
         {
             var expectedSessionId = Guid.NewGuid().ToString().Replace("-", string.Empty);
             var targetUri = new Uri("https://polydeploy.example.com/");
-            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>());
+            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>(), A.Dummy<string>());
 
             var handler = new FakeMessageHandler(
                 new Uri(targetUri, "/DesktopModules/PolyDeploy/API/Remote/CreateSession"),
@@ -38,7 +38,7 @@ namespace PolyDeploy.DeployClient.Tests
         {
             var expectedSessionId = Guid.NewGuid().ToString().Replace("-", string.Empty);
             var targetUri = new Uri("https://polydeploy.example.com/");
-            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>());
+            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>(), A.Dummy<string>());
 
             var handler = new FakeMessageHandler(
                 new Uri(targetUri, "/DesktopModules/PolyDeploy/API/Remote/CreateSession"),
@@ -55,7 +55,7 @@ namespace PolyDeploy.DeployClient.Tests
         {
             var sessionId = Guid.NewGuid().ToString().Replace("-", string.Empty);
             var targetUri = new Uri("https://polydeploy.example.com/");
-            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>());
+            var options = new DeployInput(targetUri.ToString(), A.Dummy<string>(), A.Dummy<string>());
 
             var handler = new FakeMessageHandler(
                 new Uri(targetUri, $"/DesktopModules/PolyDeploy/API/Remote/AddPackages?sessionGuid={sessionId}"),
